chore: Rename DeleteRelationshipsOfIdentityCommand to DecomposeRelationshipsOfIdentityCommand
