+++ Modules/Relationships/src/Relationships.Application/Relationships/Commands/DecomposeRelationshipsOfIdentity/DecomposeRelationshipsOfIdentityCommand.cs
@@ -1,0 +1,13 @@
+using MediatR;
+
+namespace Backbone.Modules.Relationships.Application.Relationships.Commands.DecomposeRelationshipsOfIdentity;
+
+public class DecomposeRelationshipsOfIdentityCommand : IRequest
+{
+    public DecomposeRelationshipsOfIdentityCommand(string identityAddress)
+    {
+        IdentityAddress = identityAddress;
+    }
+
+    public string IdentityAddress { get; set; }
+}
