+++ Modules/Relationships/src/Relationships.Application/Relationships/Commands/DecomposeRelationshipsOfIdentity/Handler.Tests.cs
@@ -3,23 +3,24 @@ using Backbone.Modules.Relationships.Application.Infrastructure.Persistence.Repo
 using Backbone.Modules.Relationships.Domain.Aggregates.Relationships;
 using FakeItEasy;
 
-namespace Backbone.Modules.Relationships.Application.Relationships.Commands.DeleteRelationshipsOfIdentity;
+namespace Backbone.Modules.Relationships.Application.Relationships.Commands.DecomposeRelationshipsOfIdentity;
 
 public class HandlerTests : AbstractTestsBase
 {
     [Fact]
-    public async Task Command_calls_delete_on_repository()
+    public async Task Command_calls_find_on_repository()
     {
         // Arrange
         var mockRelationshipTemplatesRepository = A.Fake<IRelationshipsRepository>();
 
         var handler = new Handler(mockRelationshipTemplatesRepository);
-        var request = new DeleteRelationshipsOfIdentityCommand(CreateRandomIdentityAddress());
+        var request = new DecomposeRelationshipsOfIdentityCommand(CreateRandomIdentityAddress());
 
         // Act
         await handler.Handle(request, CancellationToken.None);
 
         // Assert
-        A.CallTo(() => mockRelationshipTemplatesRepository.DeleteRelationships(A<Expression<Func<Relationship, bool>>>._, A<CancellationToken>._)).MustHaveHappened();
+        A.CallTo(() => mockRelationshipTemplatesRepository.FindRelationships(A<Expression<Func<Relationship, bool>>>._, A<CancellationToken>._)).MustHaveHappened();
+        A.CallTo(() => mockRelationshipTemplatesRepository.Update(A<IEnumerable<Relationship>>._)).MustHaveHappened();
     }
 }
