--- Modules/Relationships/src/Relationships.Application/Relationships/Commands/DeleteRelationshipsOfIdentity/DeleteRelationshipsOfIdentityCommand.cs
@@ -1,13 +1,0 @@
-using MediatR;
-
-namespace Backbone.Modules.Relationships.Application.Relationships.Commands.DeleteRelationshipsOfIdentity;
-
-public class DeleteRelationshipsOfIdentityCommand : IRequest
-{
-    public DeleteRelationshipsOfIdentityCommand(string identityAddress)
-    {
-        IdentityAddress = identityAddress;
-    }
-
-    public string IdentityAddress { get; set; }
-}
