--- Modules/Relationships/src/Relationships.Application/Relationships/Commands/DeleteRelationshipsOfIdentity/Validator.cs
@@ -1,13 +1,0 @@
-﻿using Backbone.BuildingBlocks.Application.Extensions;
-using Backbone.DevelopmentKit.Identity.ValueObjects;
-using FluentValidation;
-
-namespace Backbone.Modules.Relationships.Application.Relationships.Commands.DeleteRelationshipsOfIdentity;
-
-public class Validator : AbstractValidator<DeleteRelationshipsOfIdentityCommand>
-{
-    public Validator()
-    {
-        RuleFor(x => x.IdentityAddress).ValidId<DeleteRelationshipsOfIdentityCommand, IdentityAddress>();
-    }
-}
